Rename: 폴더명 변경 constant->status
